requirements.txt
fastapi
uvicorn
opencv-python
numpy
scikit-learn
python-multipart
rembg
pillow
ultralytics
onnxruntime
torch
torchvision
fuzzywuzzy
openai
flask-cors
dotenv

